refactor: rename common_specs to byte_specs, extract fuel_mono, deduplicate key decoding
