+++ rust/mcp-tools-sdk/verus-proofs/src/byte_specs.rs
