+++ rust/mcp-tools-sdk/verus-proofs/src/lib.rs
@@ -1,8 +1,9 @@
 #[allow(non_snake_case)]
-pub mod common_specs;
+pub mod byte_specs;
 pub mod decimal;
 pub mod dedup;
 pub mod escape;
+pub mod fuel_mono;
 pub mod json_spec;
 pub mod parser;
 pub mod tokenizer;
